+++ src/main/java/com/wenkrang/famara/lib/command/fa.java
@@ -1,0 +1,42 @@
+package com.wenkrang.famara.lib.command;
+
+import javax.imageio.ImageIO;
+import java.awt.*;
+import java.awt.image.BufferedImage;
+import java.io.File;
+import java.io.IOException;
+import java.util.*;
+
+public class fa {
+
+    /**
+     * 获取图片中出现最多的颜色
+     * @param imageFile PNG图片文件
+     * @return 出现最多的颜色
+     * @throws IOException 当读取图片文件失败时抛出
+     */
+    private Color getMostFrequentColor(File imageFile) throws IOException {
+        BufferedImage image = ImageIO.read(imageFile);
+        Map<Integer, Integer> colorCountMap = new HashMap<>();
+
+        // 遍历图片的每个像素
+        for (int x = 0; x < image.getWidth(); x++) {
+            for (int y = 0; y < image.getHeight(); y++) {
+                int rgb = image.getRGB(x, y);
+                colorCountMap.put(rgb, colorCountMap.getOrDefault(rgb, 0) + 1);
+            }
+        }
+
+        // 找出出现次数最多的颜色
+        int maxCount = 0;
+        int mostFrequentRGB = 0;
+        for (Map.Entry<Integer, Integer> entry : colorCountMap.entrySet()) {
+            if (entry.getValue() > maxCount) {
+                maxCount = entry.getValue();
+                mostFrequentRGB = entry.getKey();
+            }
+        }
+
+        return new Color(mostFrequentRGB);
+    }
+}
